requirements.txt
flask
pillow
numpy
imageio
scikit-image
opencv-python
easyocr
mtcnn

